feat: gathering comments
This is complicated if we want to maintain comments... but it seems like
a rare case, so we probably just get loud about how they will be
dropped/rearranged per section when rewriting to the file.

Realizing that order might be annoying to deal with if we're
file-watching... but perhaps we can parse and do an order-independent
comparison before reacting to it. An update loop comes to mind.