fix: OOM solved by removing grammar ambiguities
Adds a test case for the OOM crash, and this commit includes the
relevant comment-code for debugging this.

The solution was reducing ambiguities in whitespace matching - hopefully
this stays fast going forward.

A simple alternative that would further help is chunking the sections of
the file before parsing - the sections aren't inter-dependent anyway, so
that'd make it easier.

It'd be nice to line-by-line as well - just need to pay attention to the
multi-line dictionaries (in the inputs section). Not sure if there are
other cases like that or not (maybe for arrays/constructors), but it
seems unlikely.