refactor: dry up shared grammar parts
This might be a bad choice - it could be difficult to reason about a
grammar spread across a few variables, and maybe it's better to move
these into independent files anyway to reduce the escaping and support
remote usage/sharability...

There are a few differences between these, but fortunately some rules
can be overwritten. That might be a sign the rules aren't perfectly
architected, but meh, the tests are passing.